refactor: 알림 구조 개선
