+++ src/main/java/com/umc/banddy/domain/mypage/notification/repository/ChatNotificationRepository.java
@@ -1,7 +1,7 @@
 package com.umc.banddy.domain.mypage.notification.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.umc.banddy.domain.mypage.notification.domain.ChatNotification;
+import com.umc.banddy.domain.mypage.notification.domain.mapping.ChatNotification;
 import java.util.List;
 
 public interface ChatNotificationRepository extends JpaRepository<ChatNotification, Long> {
