+++ src/main/java/com/umc/banddy/domain/mypage/notification/repository/FriendNotificationRepository.java
@@ -1,6 +1,6 @@
 package com.umc.banddy.domain.mypage.notification.repository;
 
-import com.umc.banddy.domain.mypage.notification.domain.FriendNotification;
+import com.umc.banddy.domain.mypage.notification.domain.mapping.FriendNotification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
