+++ src/main/java/com/umc/banddy/domain/mypage/notification/service/NotificationService.java
@@ -3,9 +3,9 @@ package com.umc.banddy.domain.mypage.notification.service;
 import com.umc.banddy.domain.band.notification.domain.mapping.BandNotification;
 import com.umc.banddy.domain.band.notification.repository.BandNotificationRepository;
 import com.umc.banddy.domain.mypage.notification.converter.NotificationConverter;
-//import com.umc.banddy.domain.mypage.notification.domain.ChatNotification;
-import com.umc.banddy.domain.mypage.notification.domain.ChatNotification;
-import com.umc.banddy.domain.mypage.notification.domain.FriendNotification;
+//import com.umc.banddy.domain.mypage.notification.domain.mapping.ChatNotification;
+import com.umc.banddy.domain.mypage.notification.domain.mapping.ChatNotification;
+import com.umc.banddy.domain.mypage.notification.domain.mapping.FriendNotification;
 //import com.umc.banddy.domain.mypage.notification.repository.ChatNotificationRepository;
 import com.umc.banddy.domain.mypage.notification.repository.ChatNotificationRepository;
 import com.umc.banddy.domain.mypage.notification.repository.FriendNotificationRepository;
